Create settings menu, player class and HUD manager
This commit creates the abstract player class which contains the basic
attributes that all players in the game require. The settings menu 
allows the user to change volume of sound effects / music and
also gives the ability to switch to fullscreen mode. There is a
"back" button to return to main menu as well. A Heads Up
Display gives basic statistics that game has collected about
the human player.