Update the main menu scene
Simplify menu.py, rename it with main_menu_scene.py, and relocate it
into the scene folder.

Label(), Style(), Button() functions as well as other additional
imports are used to improve the clarity of the code and the main
menu itself.

Rebuild the code structure into a class-format and can be
called by the main controller easier.