Create Ship Class
Create a ship class that is a subclass of object. Initialize a ship,
that tracks the coordinates on the grid that taken up by this ship.
Add hit() to check if a shot at given coordinates, hits this ship.
Add is_sunk() to check for all points on the ship being hit.
Add get_hit_points() to return a list of the ships hit points.