Add Audio Files and Create an Audio Manager
The audio manager loads background music and game sound effects
so that audio can play when needed. The background music volume
can be paused, resumed and set so that user can change
it as desired in the settings menu. Sound effects like canonShot 
and shipHit are useful whenever a shot is fired or a ship is hit
respectively.